+++ fractal_server/config.py
@@ -556,6 +556,11 @@ class Settings(BaseSettings):
     Whether to include the v1 API.
     """
 
+    FRACTAL_MAX_PIP_VERSION: str = "24.0"
+    """
+    Maximum value at which to update `pip` before performing task collection.
+    """
+
     ###########################################################################
     # BUSINESS LOGIC
     ###########################################################################